back to libtiff 4.2.0

+++ libtiff/libtiff/tif_next.c
@@ -44,12 +44,12 @@
 #define WHITE   	((1<<2)-1)
 
 static int
-NeXTDecode(TIFF* tif, uint8_t* buf, tmsize_t occ, uint16_t s)
+NeXTDecode(TIFF* tif, uint8* buf, tmsize_t occ, uint16 s)
 {
 	static const char module[] = "NeXTDecode";
 	unsigned char *bp, *op;
 	tmsize_t cc;
-	uint8_t* row;
+	uint8* row;
 	tmsize_t scanline, n;
 
 	(void) s;
@@ -101,9 +101,9 @@ NeXTDecode(TIFF* tif, uint8_t* buf, tmsize_t occ, uint16_t s)
 			break;
 		}
 		default: {
-			uint32_t npixels = 0, grey;
+			uint32 npixels = 0, grey;
 			tmsize_t op_offset = 0;
-			uint32_t imagewidth = tif->tif_dir.td_imagewidth;
+			uint32 imagewidth = tif->tif_dir.td_imagewidth;
             if( isTiled(tif) )
                 imagewidth = tif->tif_dir.td_tilewidth;
 
@@ -115,7 +115,7 @@ NeXTDecode(TIFF* tif, uint8_t* buf, tmsize_t occ, uint16_t s)
 			 */
 			op = row;
 			for (;;) {
-				grey = (uint32_t)((n >> 6) & 0x3);
+				grey = (uint32)((n>>6) & 0x3);
 				n &= 0x3f;
 				/*
 				 * Ensure the run does not exceed the scanline
@@ -127,8 +127,8 @@ NeXTDecode(TIFF* tif, uint8_t* buf, tmsize_t occ, uint16_t s)
 				if (npixels >= imagewidth)
 					break;
                 if (op_offset >= scanline ) {
-                    TIFFErrorExt(tif->tif_clientdata, module, "Invalid data for scanline %"PRIu32,
-                        tif->tif_row);
+                    TIFFErrorExt(tif->tif_clientdata, module, "Invalid data for scanline %ld",
+                        (long) tif->tif_row);
                     return (0);
                 }
 				if (cc == 0)
@@ -140,17 +140,17 @@ NeXTDecode(TIFF* tif, uint8_t* buf, tmsize_t occ, uint16_t s)
 		}
 		}
 	}
-	tif->tif_rawcp = (uint8_t*) bp;
+	tif->tif_rawcp = (uint8*) bp;
 	tif->tif_rawcc = cc;
 	return (1);
 bad:
-	TIFFErrorExt(tif->tif_clientdata, module, "Not enough data for scanline %"PRIu32,
-	    tif->tif_row);
+	TIFFErrorExt(tif->tif_clientdata, module, "Not enough data for scanline %ld",
+	    (long) tif->tif_row);
 	return (0);
 }
 
 static int
-NeXTPreDecode(TIFF* tif, uint16_t s)
+NeXTPreDecode(TIFF* tif, uint16 s)
 {
 	static const char module[] = "NeXTPreDecode";
 	TIFFDirectory *td = &tif->tif_dir;
@@ -158,7 +158,7 @@ NeXTPreDecode(TIFF* tif, uint16_t s)
 
 	if( td->td_bitspersample != 2 )
 	{
-		TIFFErrorExt(tif->tif_clientdata, module, "Unsupported BitsPerSample = %"PRIu16,
+		TIFFErrorExt(tif->tif_clientdata, module, "Unsupported BitsPerSample = %d",
 					 td->td_bitspersample);
 		return (0);
 	}
